introduce kotlin-id-jvm & add first non mongo tests
