+++ build.gradle
@@ -184,11 +184,13 @@ configure(kotlinProjects) {
     group = 'org.mongodb.kotlin'
 
     dependencies {
-        implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
-        implementation "org.jetbrains.kotlin:kotlin-reflect:$kotlinVersion"
+        testImplementation(platform("org.junit:junit-bom:$junitBomVersion"))
+        testImplementation('org.junit.jupiter:junit-jupiter')
+        testImplementation('org.junit.jupiter:junit-jupiter-params')
+        testImplementation('org.junit.jupiter:junit-jupiter-engine')
     }
 
-    test{
+    test {
         useJUnitPlatform {
             includeEngines('junit-jupiter')
         }
