+++ settings.gradle
@@ -30,8 +30,8 @@ include ':driver-sync'
 include ':driver-reactive-streams'
 include ':bson-scala'
 include ':driver-scala'
-include ':driver-kotlin'
 include ':driver-kotlin:kotlin-id'
+include ':driver-kotlin:kotlin-id-jvm'
 include ':driver-kotlin:kotlin-id-jackson'
 include ':driver-kotlin:kotlin-id-serialization'
 include ':driver-kotlin:kotlin-shared'
